apps/gateway/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
greenlet>=3.0.0
sqlmodel>=0.0.22
pydantic-settings>=2.6.0
PyJWT[crypto]>=2.9.0
cryptography>=43.0.0
openai>=1.55.0
langgraph>=0.2.50
langchain-core>=0.3.20
langchain-openai>=0.2.10
aiosqlite>=0.20.0
mcp>=1.2.0
jsonschema>=4.0.0
